# swaggerDocs/swagger.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "API Documentation",
      "version": "1.0.0",
      "description": "Comprehensive API documentation."
    },
    "paths": {
      "/register": {
        "post": {
          "summary": "User Registration",
          "description": "Register a new admin user.",
          "tags": ["Authentication"],
          "security": [{ "ApiKeyAuth": [] }],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_email": {
                      "type": "string",
                      "example": "test@test.com"
                    },
                    "user_password": {
                      "type": "string",
                      "example": "password"
                    }
                  },
                  "required": ["user_email", "user_password"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "User registered successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "boolean",
                        "example": true
                      },
                      "message": {
                        "type": "string",
                        "example": "User registered successfully"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "User already exists",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "boolean",
                        "example": false
                      },
                      "message": {
                        "type": "string",
                        "example": "User already exists"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/login": {
        "post": {
          "summary": "User Login",
          "description": "Authenticate an admin user and generate a JWT token.",
          "tags": ["Authentication"],
          "security": [{ "ApiKeyAuth": [] }],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_email": {
                      "type": "string",
                      "example": "random@org.org"
                    },
                    "user_password": {
                      "type": "string",
                      "example": "randomFun"
                    }
                  },
                  "required": ["user_email", "user_password"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful login",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "boolean",
                        "example": true
                      },
                      "message": {
                        "type": "string",
                        "example": "User logged in successfully"
                      },
                      "token": {
                        "type": "string",
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2VtYWlsIjoicmFuZG9tQG9yZy5vcmciLCJ1c2VyX2lkIjoxLCJ1c2VyX2xldmVsIjoiU1VQRVJfQURNSU4iLCJpYXQiOjE3MzQwODk1NDUsImV4cCI6MTczNDA5MzE0NX0.ej9hLIQdXJtHVCvs87Zn3QTHH7uaR6YaX2RYylIsWgk"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid email or password",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "boolean",
                        "example": false
                      },
                      "message": {
                        "type": "string",
                        "example": "Invalid email or password"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/logout": {
        "post": {
          "summary": "User Logout",
          "description": "Logout the currently authenticated user.",
          "tags": ["Authentication"],
          "security": [{ "ApiKeyAuth": [] }],
          "parameters": [
            {
              "name": "secret",
              "in": "header",
              "required": true,
              "schema": {
                "type": "string"
              },
              "example": "<token>"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful logout",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "boolean",
                        "example": true
                      },
                      "message": {
                        "type": "string",
                        "example": "User logged out successfully"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/validateToken": {
        "post": {
          "summary": "Validate Token",
          "description": "Validate the provided JWT token.",
          "tags": ["Authentication"],
          "security": [{ "ApiKeyAuth": [] }],
          "parameters": [
            {
              "name": "secret",
              "in": "header",
              "required": true,
              "schema": {
                "type": "string"
              },
              "example": "<token>"
            }
          ],
          "responses": {
            "200": {
              "description": "Token is valid",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "boolean",
                        "example": true
                      },
                      "message": {
                        "type": "string",
                        "example": "Token is valid"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Invalid or expired token",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "boolean",
                        "example": false
                      },
                      "message": {
                        "type": "string",
                        "example": "Token is invalid or expired"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "/getClanInfo/{tag}": {
        "post": {
          "tags": ["Clan Endpoints"],
          "summary": "Get information about a clan",
          "parameters": [
            {
              "name": "tag",
              "in": "path",
              "required": true,
              "description": "The tag of the clan",
              "schema": {
                "type": "string"
              }
            }
          ],
          "security": [{ "ApiKeyAuth": [] }],
          "responses": {
            "200": {
              "description": "Clan data retrieved successfully"
            }
          }
        }
      },
      "/getClanMembers/{tag}": {
        "post": {
          "tags": ["Clan Endpoints"],
          "summary": "Get members of a clan",
          "parameters": [
            {
              "name": "tag",
              "in": "path",
              "required": true,
              "description": "The tag of the clan",
              "schema": {
                "type": "string"
              }
            }
          ],
          "security": [{ "ApiKeyAuth": [] }],
          "responses": {
            "200": {
              "description": "Clan members retrieved successfully"
            }
          }
        }
      },
      "/getCapitalRaidSeasons/{tag}": {
        "post": {
          "tags": ["Clan Endpoints"],
          "summary": "Get capital raid seasons for a clan",
          "parameters": [
            {
              "name": "tag",
              "in": "path",
              "required": true,
              "description": "The tag of the clan",
              "schema": {
                "type": "string"
              }
            }
          ],
          "security": [{ "ApiKeyAuth": [] }],
          "responses": {
            "200": {
              "description": "Capital raid seasons data retrieved successfully"
            }
          }
        }
      },
      "/getClanMembersHero/{tag}": {
        "post": {
          "tags": ["Clan Endpoints"],
          "summary": "Get hero data for clan members",
          "parameters": [
            {
              "name": "tag",
              "in": "path",
              "required": true,
              "description": "The tag of the clan",
              "schema": {
                "type": "string"
              }
            }
          ],
          "security": [{ "ApiKeyAuth": [] }],
          "responses": {
            "200": {
              "description": "Hero data retrieved successfully"
            }
          }
        }
      },
      "/getPlayersInfo/{tag}": {
        "post": {
          "tags": ["Player Endpoints"],
          "summary": "Get player information",
          "parameters": [
            {
              "name": "tag",
              "in": "path",
              "required": true,
              "description": "The tag of the player",
              "schema": {
                "type": "string"
              }
            }
          ],
          "security": [{ "ApiKeyAuth": [] }],
          "responses": {
            "200": {
              "description": "Player information retrieved successfully"
            }
          }
        }
      },
      "/getCurrentWar/{tag}": {
        "post": {
          "tags": ["War Endpoints"],
          "summary": "Get the current war details",
          "parameters": [
            {
              "name": "tag",
              "in": "path",
              "required": true,
              "description": "The tag of the clan",
              "schema": {
                "type": "string"
              }
            }
          ],
          "security": [{ "ApiKeyAuth": [] }],
          "responses": {
            "200": {
              "description": "Current war details retrieved successfully"
            }
          }
        }
      },
      "/getWarLog/{tag}": {
        "post": {
          "tags": ["War Endpoints"],
          "summary": "Get the war log for a clan",
          "parameters": [
            {
              "name": "tag",
              "in": "path",
              "required": true,
              "description": "The tag of the clan",
              "schema": {
                "type": "string"
              }
            }
          ],
          "security": [{ "ApiKeyAuth": [] }],
          "responses": {
            "200": {
              "description": "War log retrieved successfully"
            }
          }
        }
      },
      "/getTHLevels/{tag}": {
        "post": {
          "tags": ["Clan Endpoints"],
          "summary": "Get Town Hall levels for clan members",
          "parameters": [
            {
              "name": "tag",
              "in": "path",
              "required": true,
              "description": "The tag of the clan",
              "schema": {
                "type": "string"
              }
            }
          ],
          "security": [{ "ApiKeyAuth": [] }],
          "responses": {
            "200": {
              "description": "Town Hall levels retrieved successfully"
            }
          }
        }
      },
      "/getCWLresults/{tag}": {
        "post": {
          "tags": ["War Endpoints"],
          "summary": "Get Clan War League results",
          "parameters": [
            {
              "name": "tag",
              "in": "path",
              "required": true,
              "description": "The tag of the clan",
              "schema": {
                "type": "string"
              }
            }
          ],
          "security": [{ "ApiKeyAuth": [] }],
          "responses": {
            "200": {
              "description": "CWL results retrieved successfully"
            }
          }
        }
      },
      "/getWarWeight/{tag}": {
        "post": {
          "tags": ["War Endpoints"],
          "summary": "Get the war weight of a clan",
          "parameters": [
            {
              "name": "tag",
              "in": "path",
              "required": true,
              "description": "The tag of the clan",
              "schema": {
                "type": "string"
              }
            }
          ],
          "security": [{ "ApiKeyAuth": [] }],
          "responses": {
            "200": {
              "description": "War weight retrieved successfully"
            }
          }
        }
      },
      "/db/getRecords/{phoneNumber}": {
        "post": {
          "tags": ["Database Operations"],
          "summary": "Get database records by phone number",
          "parameters": [
            {
              "name": "phoneNumber",
              "in": "path",
              "required": true,
              "description": "Phone number to query records",
              "schema": {
                "type": "string"
              }
            }
          ],
          "security": [{ "ApiKeyAuth": [] }],
          "responses": {
            "200": {
              "description": "Records retrieved successfully"
            }
          }
        }
      },
      "/db/addRecord/{PhoneNumber}/{Type}/{Tag}/{Action}": {
        "post": {
          "tags": ["Database Operations"],
          "summary": "Add a database record",
          "parameters": [
            {
              "name": "PhoneNumber",
              "in": "path",
              "required": true,
              "description": "Phone number for the record",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "Type",
              "in": "path",
              "required": true,
              "description": "Record type",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "Tag",
              "in": "path",
              "required": true,
              "description": "Record tag",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "Action",
              "in": "path",
              "required": true,
              "description": "Action performed",
              "schema": {
                "type": "string"
              }
            }
          ],
          "security": [{ "ApiKeyAuth": [] }],
          "responses": {
            "200": {
              "description": "Record added successfully"
            }
          }
        }
      },
      "/getPlayersHistory/{tag}": {
        "post": {
          "tags": ["Player Endpoints"],
          "summary": "Get the history of a player",
          "parameters": [
            {
              "name": "tag",
              "in": "path",
              "required": true,
              "description": "The tag of the player",
              "schema": {
                "type": "string"
              }
            }
          ],
          "security": [{ "ApiKeyAuth": [] }],
          "responses": {
            "200": {
              "description": "Player history retrieved successfully"
            }
          }
        }
      },
      "/getClanRankingsFromLocation/{locationId}": {
      "post": {
        "summary": "Get Clan Rankings from Location",
        "description": "Fetches the rankings of clans from a given location.",
        "tags": ["Location Endpoints"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "description": "ID of the location to get clan rankings from",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Limit the number of results returned. Defaults to 10 if not specified.",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with clan rankings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "clanId": { "type": "string" },
                          "clanName": { "type": "string" },
                          "rank": { "type": "integer" },
                          "members": { "type": "integer" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/getPlayersRankingsFromLocation/{locationId}": {
      "post": {
        "summary": "Get Player Rankings from Location",
        "description": "Fetches the rankings of players from a given location.",
        "tags": ["Location Endpoints"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "description": "ID of the location to get player rankings from",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Limit the number of results returned. Defaults to 10 if not specified.",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with player rankings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "playerId": { "type": "string" },
                          "playerName": { "type": "string" },
                          "rank": { "type": "integer" },
                          "level": { "type": "integer" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/get/ClanAssociation/{tag}": {
      "post": {
        "summary": "Get Clan Association",
        "description": "Retrieve clan association details by clan tag.",
        "tags": ["Association"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "name": "tag",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Clan tag to fetch association details for."
          }
        ],
        "responses": {
          "200": {
            "description": "Clan association details fetched successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "clan_id": { "type": "integer" },
                        "majorLeagueInfo": { "type": "string" },
                        "minorLeagueInfo": { "type": "string" },
                        "publicNote": { "type": "string" },
                        "internalNote": { "type": "string" },
                        "last_updated": { "type": "string", "format": "date-time" },
                        "isSynced": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/add/ClanAssociation/{tag}": {
      "post": {
        "summary": "Add Clan Association",
        "description": "Add or update clan association details by clan tag.",
        "tags": ["Association"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "name": "tag",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Clan tag to add or update association details for."
          },
          {
            "name": "secret",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "<token>"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "majorLeagueInfo": { "type": "string", "default": "" },
                  "minorLeagueInfo": { "type": "string", "default": "" },
                  "publicNote": { "type": "string", "default": "" },
                  "internalNote": { "type": "string", "default": "" }
                },
                "required": ["majorLeagueInfo", "minorLeagueInfo", "publicNote", "internalNote"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Clan association details added or updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "clan_id": { "type": "integer" },
                        "majorLeagueInfo": { "type": "string" },
                        "minorLeagueInfo": { "type": "string" },
                        "publicNote": { "type": "string" },
                        "internalNote": { "type": "string" },
                        "last_updated": { "type": "string", "format": "date-time" },
                        "isSynced": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
...
      "/getTrackedClanWars/{tag}": {
        "post": {
          "tags": ["War Endpoints"],
          "summary": "Get tracked clan wars",
          "parameters": [
            {
              "name": "tag",
              "in": "path",
              "required": true,
              "description": "The tag of the clan",
              "schema": {
                "type": "string"
              }
            }
          ],
          "security": [{ "ApiKeyAuth": [] }],
          "responses": {
            "200": {
              "description": "Tracked clan wars retrieved successfully"
            }
          }
        }
      },
...
    },
    "components": {
      "securitySchemes": {
        "ApiKeyAuth": {
          "type": "apiKey",
          "name": "Authorization",
          "in": "header"
        }
      }
    }
  }